Last commit

+++ src/main/java/Nhom09_WebNgheNhac/Nhom09_WebNgheNhac/Controller/SongController.java
@@ -1,17 +1,24 @@
 package Nhom09_WebNgheNhac.Nhom09_WebNgheNhac.Controller;
 
 import Nhom09_WebNgheNhac.Nhom09_WebNgheNhac.Model.Category;
+import Nhom09_WebNgheNhac.Nhom09_WebNgheNhac.Model.Playlist;
 import Nhom09_WebNgheNhac.Nhom09_WebNgheNhac.Model.Song;
+import Nhom09_WebNgheNhac.Nhom09_WebNgheNhac.Model.User;
+import Nhom09_WebNgheNhac.Nhom09_WebNgheNhac.Role;
 import Nhom09_WebNgheNhac.Nhom09_WebNgheNhac.Service.CategoryService;
-import Nhom09_WebNgheNhac.Nhom09_WebNgheNhac.Service.SingerService;
+import Nhom09_WebNgheNhac.Nhom09_WebNgheNhac.Service.PlaylistService;
 import Nhom09_WebNgheNhac.Nhom09_WebNgheNhac.Service.SongService;
 import Nhom09_WebNgheNhac.Nhom09_WebNgheNhac.Service.UserService;
 import com.mpatric.mp3agic.InvalidDataException;
 import com.mpatric.mp3agic.UnsupportedTagException;
 import io.micrometer.common.lang.NonNull;
 import jakarta.validation.Valid;
-import jakarta.validation.constraints.NotNull;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.access.AccessDeniedException;
+import org.springframework.security.authentication.AnonymousAuthenticationToken;
+import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.validation.BindingResult;
@@ -19,8 +26,7 @@ import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.io.IOException;
-import java.util.List;
-import java.util.Optional;
+import java.util.*;
 import java.util.stream.Collectors;
 
 @Controller
@@ -31,65 +37,169 @@ public class SongController {
     @Autowired
     private CategoryService categoryService;
     @Autowired
-    private SingerService singerService;
+    private UserService userService;
+    @Autowired
+    private PlaylistService playlistService;
+
 
     @GetMapping
-    public String listSong(Model model) {
-        model.addAttribute("songs", songService.getAllSong());
+    public String listSong(Model model , String sapxep) {
+        if(sapxep!=null){
+            if(sapxep.equals("tangdan")){
+                model.addAttribute("songs", songService.getAllSong().stream().filter(s -> !s.isDelete()).sorted(Comparator.comparing(Song::getLikeCount)).toList());
+            }
+            else
+                model.addAttribute("songs", songService.getAllSong().stream().filter(s -> !s.isDelete()).sorted(Comparator.comparing(Song::getLikeCount).reversed()).toList());
+        }
+        else
+            model.addAttribute("songs", songService.getAllSong().stream().filter(s -> !s.isDelete()).toList());
+
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        List<Integer> songIds = new ArrayList<>();
+        if (!(authentication instanceof AnonymousAuthenticationToken)) {
+            User user = (User) authentication.getPrincipal();
+            Optional<User> user1 = userService.getUserId(user.getUserId());
+            UsernamePasswordAuthenticationToken newAuthentication = new UsernamePasswordAuthenticationToken(user1.get(), authentication.getCredentials(), user1.get().getAuthorities());
+            SecurityContextHolder.getContext().setAuthentication(newAuthentication);
+
+
+            Playlist playlist = playlistService.likePlaylist(user.getUserId(), 1);
+
+            songIds = playlist.getSongPlaylist()
+                    .stream()
+                    .map(Song::getSongId)
+                    .toList();
+
+            List<Playlist> playlists = playlistService.getPlaylistsByUser(user).stream().filter(p->p.getCategoryPlaylist().getCategoryPlaylistId()!=1).toList();
+            
+            model.addAttribute("playlists", playlists);
+
+        }
+
+        List<Category> category = categoryService.getAlCatologies();
+        model.addAttribute("category", category);
+        model.addAttribute("songIds", songIds);
         return "/song/list-song";
     }
 
     @GetMapping("/song/add")
     public String showAddForm(Model model) {
         model.addAttribute("song", new Song());
-        model.addAttribute("categories", categoryService.getAlCatologies());
-        model.addAttribute("singers", singerService.getAllSinger());
+        model.addAttribute("categories", categoryService.getAlCatologies()
+                                                        .stream().filter(c -> !c.isDelete()).toList());
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        User user = (User) authentication.getPrincipal();
+        List<User> singers = userService.getAllUser().stream()
+                .filter(p -> !p.getUserId().equals(user.getUserId()))
+                .filter(p -> p.getRoles().stream().anyMatch(role -> role.getRoleName().equals("SINGER")))
+                .collect(Collectors.toList());
+        model.addAttribute("singers", singers);
+
         return "/song/add-song";
     }
 
     @PostMapping("/song/add")
-    public String addProduct(@Valid Song song, BindingResult result,MultipartFile imageFile,MultipartFile fileMp3,Model model) throws InvalidDataException, UnsupportedTagException, IOException {
+    public String addProduct(@Valid Song song, BindingResult result, MultipartFile imageFile, MultipartFile fileMp3,
+            String selectedNgheSiList, Model model) throws InvalidDataException, UnsupportedTagException, IOException {
         if (result.hasErrors()) {
-            model.addAttribute("categories", categoryService.getAlCatologies());
+            model.addAttribute("categories", categoryService.getAlCatologies()
+                                                .stream().filter(c -> !c.isDelete()).toList());
             return "/song/add-song";
         }
         song.setTime(songService.timeMusic(fileMp3));
         song.setImage(songService.saveImage(imageFile));
         song.setFilePath(songService.saveMusic(fileMp3));
-        songService.addSong(song);
+        songService.addSong(song, selectedNgheSiList);
         return "redirect:/";
     }
 
+    @GetMapping("/song/detail/{songId}")
+    public String detailSong(@PathVariable("songId") int songId, Model model) {
+
+        Song song = songService.getSongId(songId)
+                .orElseThrow(() -> new IllegalArgumentException("Invalid song Id:" + songId));
+
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        List<Integer> songIds = new ArrayList<>();
+        boolean check;
+
+        if(!(authentication instanceof AnonymousAuthenticationToken)){
+            User user = (User) authentication.getPrincipal();
+            Playlist playlist = playlistService.likePlaylist(user.getUserId(), 1);
+            List<Playlist> playlists = playlistService.getPlaylistsByUser(user).stream().filter(p->p.getCategoryPlaylist().getCategoryPlaylistId()==2).toList();
+            model.addAttribute("playlists", playlists);
+            songIds = playlist.getSongPlaylist()
+                    .stream()
+                    .map(Song::getSongId)
+                    .toList();
+
+            if(!user.getUserId().equals(song.getCreateByUser()) && !user.getRoles().stream()
+                    .anyMatch(p -> p.getRoleId().equals(Role.ADMIN.value)) ){
+                if(song.isPremium()){
+                    if(user.isPremium())
+                        check= true;
+                    else
+                        check = false;
+                }
+                else
+                    check = true;
+            }
+            else
+                check = true;
+        }
+        else{
+            if(song.isPremium()){
+                check = false;
+            }
+            else
+                check = true;
+        }
+        String userName = songService.getSongId(songId).get().getUsers().stream().map(User ::getFullName)
+                .collect(Collectors.joining(", "));
+
+        model.addAttribute("userName", userName);
+        model.addAttribute("songIds", songIds);
+        model.addAttribute("check", check);
+        model.addAttribute("song", song);
+        return "/song/detail-song";
+    }
 
     @GetMapping("/song/edit/{songId}")
     public String showUpdateForm(@PathVariable("songId") int songId, Model model) {
         Song song = songService.getSongId(songId)
                 .orElseThrow(() -> new IllegalArgumentException("Invalid song Id:" + songId));
-        model.addAttribute("categories", categoryService.getAlCatologies());
-        model.addAttribute("singers", singerService.getAllSinger());
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        User userLogin = (User) authentication.getPrincipal();
+        if (!song.getCreateByUser().equals(userLogin.getUserId())) {
+            return "redirect:/error";
+        }
+        model.addAttribute("categories", categoryService.getAlCatologies()
+                .stream().filter(c -> !c.isDelete()).toList());
         model.addAttribute("song", song);
         return "/song/update-song";
     }
 
     @PostMapping("/song/edit/{songId}")
-    public String updateProduct(@PathVariable("songId") int songId, @Valid Song song, BindingResult result,MultipartFile imageFile,MultipartFile fileMp3) throws IOException, InvalidDataException, UnsupportedTagException {
+    public String updateProduct(@PathVariable("songId") int songId, @Valid Song song, BindingResult result,
+            MultipartFile imageFile, MultipartFile fileMp3, Model model)
+            throws IOException, InvalidDataException, UnsupportedTagException {
         if (result.hasErrors()) {
-            song.setSongId(Math.toIntExact(songId));
-            return "/products/update-product";
+            model.addAttribute("categories", categoryService.getAlCatologies()
+                    .stream().filter(c -> !c.isDelete()).toList());
+            return "/song/update-song";
         }
 
         Optional<Song> song1 = songService.getSongId(songId);
-        if(!imageFile.isEmpty())
+        if (!imageFile.isEmpty())
             song.setImage(songService.saveImage(imageFile));
-        else{
+        else {
             song.setImage(song1.get().getImage());
         }
 
-        if(!fileMp3.isEmpty()){
+        if (!fileMp3.isEmpty()) {
             song.setTime(songService.timeMusic(fileMp3));
             song.setFilePath(songService.saveMusic(fileMp3));
-        }
-        else{
+        } else {
             song.setTime(song1.get().getTime());
             song.setFilePath(song1.get().getFilePath());
         }
@@ -100,24 +210,95 @@ public class SongController {
     }
 
     @GetMapping("/song/delete/{songId}")
-    public String deleteCategory(@PathVariable("songId") int songId, Model model) {
+    public String deleteSong(@PathVariable("songId") int songId, Model model) {
         Song song = songService.getSongId(songId)
                 .orElseThrow(() -> new IllegalArgumentException("Invalid song Id:" + songId));
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        User userLogin = (User) authentication.getPrincipal();
+        if (!song.getCreateByUser().equals(userLogin.getUserId())) {
+            return "redirect:/error";
+        }
         songService.deleteSongById(songId);
         return "redirect:/";
     }
 
 
     @GetMapping("/search")
-    public String Search(@NonNull Model model, String query) {
+    public String Search(@NonNull Model model, String query , String sapxep) {
+        if(sapxep!=null){
+            if(sapxep.equals("tangdan")){
+                model.addAttribute("songs", songService.searchSong(query).stream().sorted(Comparator.comparing(Song::getLikeCount)));
+            }
+            else
+                model.addAttribute("songs", songService.searchSong(query).stream().sorted(Comparator.comparing(Song::getLikeCount).reversed()));
+        }
+        else
+            model.addAttribute("songs", songService.searchSong(query));
 
-        List<Song> songs = songService.getAllSong();
-        model.addAttribute("songs", songs.stream()
-                .filter(title -> title.getSongName().toLowerCase().contains(query.toLowerCase()))
-                .collect(Collectors.toList()));
+        model.addAttribute("query", query);
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        List<Integer> songIds = new ArrayList<>();
+        if(!(authentication instanceof AnonymousAuthenticationToken)){
+            User user = (User) authentication.getPrincipal();
+
+            Playlist playlist = playlistService.likePlaylist(user.getUserId(),1);
+
+            songIds = playlist.getSongPlaylist()
+                    .stream()
+                    .map(Song::getSongId)
+                    .toList();
+        }
+        List<Category> category = categoryService.getAlCatologies();
+        model.addAttribute("category", category);
+        model.addAttribute("songIds", songIds);
         return "/song/list-song";
     }
 
+    @GetMapping("/SearchSuggestions")
+    @ResponseBody
+    public List<String> searchSuggestions(String query) {
+        return songService.SearchSuggestions(query);
+    }
 
+
+    @GetMapping("/song/manage")
+    public String listQuanLi(Model model) {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        User user = (User) authentication.getPrincipal();
+        List<Playlist> playlists = playlistService.getPlaylistsByUser(user).stream().filter(p->p.getCategoryPlaylist().getCategoryPlaylistId()==3).toList();
+        model.addAttribute("playlists", playlists);
+        model.addAttribute("songs",
+                songService.getAllSong().stream().filter(p -> p.getCreateByUser().equals(user.getUserId()) && !p.isDelete()));
+        return "/song/manage-song";
+    }
+
+    @GetMapping("/song/like/{songId}")
+    public String likeSong(@PathVariable("songId") int songId, Model model) {
+        Song song = songService.getSongId(songId).stream().findFirst()
+                .orElseThrow(() -> new IllegalArgumentException("Invalid song Id:" + songId));
+
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        User user = (User) authentication.getPrincipal();
+        Playlist playlist = playlistService.likePlaylist(user.getUserId(), 1);
+        Set<Song> listSong = playlist.getSongPlaylist();
+        if(listSong.stream().anyMatch(p->p.getSongId() == song.getSongId())){
+            listSong.removeIf(p ->p.getSongId() == song.getSongId());
+            song.setLikeCount(song.getLikeCount()-1);
+        }
+        else{
+            listSong.add(song);
+            song.setLikeCount(song.getLikeCount()+1);
+        }
+
+        songService.updateSong(song);
+        playlist.setSongPlaylist(listSong);
+        playlistService.update(playlist);
+        return "redirect:/song/detail/" + songId;
+    }
+
+    @GetMapping("/error")
+    public String error(){
+        return "/error";
+    }
 
 }
